PM-33:feat:role删除 补充逻辑，被引用的Role不允许删除

+++ pm-application/src/main/java/com/pm/application/service/impl/RoleServiceImpl.java
@@ -8,10 +8,16 @@ import com.pm.application.convertor.RoleConvertor;
 import com.pm.application.dto.cmd.RoleAddCmd;
 import com.pm.application.dto.vo.RoleVO;
 import com.pm.application.service.IRoleService;
+import com.pm.infrastructure.consts.ErrorCodeEnum;
 import com.pm.infrastructure.dataobject.RoleDO;
+import com.pm.infrastructure.dataobject.RoleResourceDO;
+import com.pm.infrastructure.dataobject.UserRoleDO;
 import com.pm.infrastructure.entity.PageQuery;
 import com.pm.infrastructure.entity.PageResponse;
 import com.pm.infrastructure.mapper.RoleMapper;
+import com.pm.infrastructure.mapper.RoleResourceMapper;
+import com.pm.infrastructure.mapper.UserRoleMapper;
+import com.zyzh.exception.BizException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -26,6 +32,12 @@ public class RoleServiceImpl implements IRoleService {
 
     @Autowired
     private RoleMapper roleMapper;
+
+    @Autowired
+    private UserRoleMapper userRoleMapper;
+
+    @Autowired
+    private RoleResourceMapper roleResourceMapper;
 
     @Override
     public PageResponse<RoleVO> pageRole(PageQuery pageQuery) {
@@ -49,6 +61,11 @@ public class RoleServiceImpl implements IRoleService {
 
     @Override
     public Response deleteRole(String id) {
+        Integer userRoleCount = userRoleMapper.selectCount(new LambdaQueryWrapper<UserRoleDO>().eq(UserRoleDO::getRoleId, id));
+        Integer roleResourceCount = roleResourceMapper.selectCount(new LambdaQueryWrapper<RoleResourceDO>().eq(RoleResourceDO::getRoleId, id));
+        if (userRoleCount > 0 || roleResourceCount > 0) {
+            throw new BizException(ErrorCodeEnum.ROLE_NOT_ALLOW_DELETE);
+        }
         roleMapper.deleteById(id);
         return Response.buildSuccess();
     }
